refactor(voice): extract TTSSynthesisQueue to tts_synthesis.py (issue #94 phase 3)
把「合成队列」职责从 GPTSoVITSTTSProvider 抽到新模块 tts_synthesis.py：
- TTSSynthesisQueue：串行化 speak/prepare 请求，后台 daemon 线程合成，结果交回
  播放端点（sink）；保持「每请求一次性 daemon 线程 + _request_running 串行」线程模型
- 合成线程登记进协调器自持 RM 的 ThreadResource，关闭随 stop_all 走 cancel→join→linger
- GPT/Genie 合成差异封装为 GPTSoVITSSynthesisEngine / GenieSynthesisEngine，
  队列只调度；GenieTTSProvider 不再覆写 _request_audio，仅 _create_supervisor +
  _create_synthesis_engine 装配选型
- 协调器实现 TTSSynthesisSink（deliver_audio/deliver_prepared/fail/skip/schedule_cleanup），
  以 self 为 sink；speak/prepare 委托 submit，discard_prepared 委托队列 + 清播放队列
- _closed 改由独立 _close_lock 守护，避免与队列锁形成反向锁序（submit/_start_next
  在持队列锁前检查 is_closed）
- 迁 _is_voiceable_text/_resolve_request_text_lang/_write_genie_audio 等合成 helper 与
  文本常量；tts.py re-export 供既有导入
- 重写 close 队列测试指向 _synthesis_queue

回归：tests/unit + tests/ui 全绿（988 passed，排除 2 个已知环境性失败）。
